requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
